start factoring out scoring module
